add basic support for the creation of ctfEvents
Note: there is a small bug present with the datetime pickers. It
seems that angular doesn't sense a change to the input field until
the user manually applies a change. So for now just delete and
re-add the Z at the end of the input string. (You can test this
out by viewing the edit form for an event. I left a debug header
that corresponds to the start date field)